refactor: move AddVariableButton to Footer directory
